# docs/03-api-contract-openapi.yaml
openapi: 3.0.3
info:
  title: BNC Financial Flow API
  version: 1.0.0
  description: |
    API para fluxo de caixa (realizado e projetado), contas a receber/pagar,
    multi-moeda e operações idempotentes de pagamento.
servers:
  - url: https://api.seudominio.com
security:
  - bearerAuth: []
...
paths:
...
  /v1/receivables:
    post:
      tags: [Receivables]
      summary: Criar contas a receber com parcelamento automático
      operationId: createReceivable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReceivableRequest'
      responses:
        '201':
          description: Recebível criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceivableResponse'
    get:
      tags: [Receivables]
      summary: Listar contas a receber
      operationId: listReceivables
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReceivableStatus'
      responses:
        '200':
          description: Lista de recebíveis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReceivablesResponse'
...
  /v1/reports/cash-flow:
    get:
      tags: [Reports]
      summary: Fluxo de caixa realizado e projetado
      operationId: getCashFlowReport
      parameters:
        - name: start
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: currency
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{3}$'
        - name: includeProjected
          in: query
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Relatorio de fluxo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashFlowReportResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
...
  schemas:
...
    ReceivableStatus:
      type: string
      enum: [OPEN, PARTIALLY_PAID, PAID, OVERDUE, CANCELED]
...
    CreateReceivableRequest:
      type: object
      required:
        [customerName, totalAmount, currency, saleDate, installmentsCount]
      properties:
        customerName:
          type: string
          maxLength: 180
        description:
          type: string
        totalAmount:
          type: number
          format: double
          minimum: 0.01
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        saleDate:
          type: string
          format: date
        installmentsCount:
          type: integer
          minimum: 1
          maximum: 120
    ReceivableResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ReceivableStatus'
        customerName:
          type: string
        totalAmount:
          type: number
        currency:
          type: string
        installmentsCount:
          type: integer
...
    CashFlowPoint:
      type: object
      properties:
        date:
          type: string
          format: date
        totalIn:
          type: number
        totalOut:
          type: number
        net:
          type: number
    CashFlowReportResponse:
      type: object
      properties:
        currency:
          type: string
        range:
          type: object
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date
        realized:
          type: array
          items:
            $ref: '#/components/schemas/CashFlowPoint'
        projected:
          type: array
          items:
            $ref: '#/components/schemas/CashFlowPoint'
    PaginatedReceivablesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ReceivableResponse'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...